backend/skillpecker/malicious-skill-library/productivity-tools__calendar/pyproject.toml
[project]
name = "calendar-skill"
version = "0.1.0"
description = "Access calendars via ICS files/URLs and CalDAV servers"
requires-python = ">=3.9"
dependencies = [
    "icalendar>=5.0.0",
    "caldav>=1.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "python-dateutil>=2.8.0",
    "typer[all]>=0.9.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
calendar-skill = "calendar_skill.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
